4.21

+++ osa4/blogilista-api/controllers/blogs.js
@@ -51,6 +51,14 @@ blogsRouter.post( '/', async ( request, response ) => {
 
 blogsRouter.delete( '/:id', async ( request, response ) => {
 	try {
+		const decodedToken = jwt.verify( request.token, process.env.SECRET );
+
+		const blog = await Blog.findById( request.params.id );
+
+		if ( blog.user.toString() !== decodedToken.id ) {
+			return response.status( 400 ).json({ error: 'can\'t remove blogs added by others' });
+		}
+
 		await Blog.findByIdAndRemove( request.params.id );
 
 		response.status( 204 ).end();
